+++ com.unity.cinemachine/Runtime/Core/ConfinerOven.cs
@@ -46,7 +46,8 @@ namespace Unity.Cinemachine
 
             public Vector2 ConfinePoint(in Vector2 pointToConfine)
             {
-                if (m_Solution.Count <= 0) return pointToConfine; // empty confiner -> no need to confine
+                if (m_Solution.Count <= 0) 
+                    return pointToConfine; // empty confiner -> no need to confine
 
                 Vector2 pInConfinerSpace = m_AspectStretcher.Stretch(pointToConfine);
                 var p = new IntPoint(pInConfinerSpace.x * k_FloatToIntScaler, pInConfinerSpace.y * k_FloatToIntScaler);
@@ -134,8 +135,9 @@ namespace Unity.Cinemachine
                 
                 bool DoesIntersectOriginal(IntPoint l1, IntPoint l2)
                 {
-                    foreach (var original in m_OriginalPolygon)
+                    for (int p = 0; p < m_OriginalPolygon.Count; ++p)
                     {
+                        var original = m_OriginalPolygon[p];
                         var numPoints = original.Count;
                         for (var i = 0; i < numPoints; ++i)
                             if (FindIntersection(l1, l2, original[i], original[(i + 1) % numPoints]) == 2)