+++ com.unity.cinemachine/Runtime/Core/InputAxisControllerBase.cs
@@ -35,6 +35,7 @@ namespace Unity.Cinemachine
     [Serializable]
     internal class InputAxisControllerManager<T> where T : IInputAxisReader, new ()
     {
+        [NonReorderable]
         public List<InputAxisControllerBase<T>.Controller> Controllers = new ();
 
         /// Axes are dynamically discovered by querying behaviours implementing <see cref="IInputAxisOwner"/>
@@ -122,8 +123,9 @@ namespace Unity.Cinemachine
 
             // Rebuild the controller list, recycling existing ones to preserve the settings
             List<InputAxisControllerBase<T>.Controller> newControllers = new();
-            foreach (var t in m_AxisOwners)
+            for (int j = 0; j < m_AxisOwners.Count; ++j)
             {
+                var t = m_AxisOwners[j];
                 var startIndex = m_Axes.Count;
                 t.GetInputAxes(m_Axes);
                 for (int i = startIndex; i < m_Axes.Count; ++i)