+++ com.unity.cinemachine/Samples~/Shared Assets/Scripts/SimplePlayerController.cs
@@ -5,25 +5,17 @@ using UnityEngine.Events;
 
 namespace Unity.Cinemachine.Samples
 {
-    public class SimplePlayerController : MonoBehaviour, IInputAxisOwner
+    public abstract class SimplePlayerControllerBase : MonoBehaviour, IInputAxisOwner
     {
         public float Speed = 1f;
         public float SprintSpeed = 4;
         public float JumpSpeed = 4;
         public float SprintJumpSpeed = 6;
     
-        public float Damping = 0.5f;
-
-        public enum ForwardModes { Camera, Player, World };
-        public ForwardModes InputForward = ForwardModes.Camera;
-        public enum UpModes { Player, World };
-        public UpModes UpMode = UpModes.World;
-
-        public bool Strafe = false;
         public bool LockCursor = false;
 
         public Action PreUpdate;
-        public Action PostUpdate;
+        public Action<Vector3, float> PostUpdate;
         public Action StartJump;
         public Action EndJump;
 
@@ -43,26 +35,7 @@ namespace Unity.Cinemachine.Samples
         [Header("Events")]
         [Tooltip("This event is sent when the player lands after a jump.")]
         public UnityEvent Landed = new ();
-
-        Vector3 m_CurrentVelocityXZ;
-        Vector3 m_LastInput;
-        float m_CurrentVelocityY;
-        bool m_IsSprinting;
-        bool m_IsJumping;
-        CharacterController m_Controller; // optional
-
-        // Get velocity relative to player transform
-        public Vector3 GetPlayerVelocity()
-        {
-            var vel = Quaternion.Inverse(transform.rotation) * m_CurrentVelocityXZ;
-            vel.y = m_CurrentVelocityY;
-            return vel;
-        }
-
-        public bool IsSprinting => m_IsSprinting;
-        public bool IsJumping => m_IsJumping;
-        public bool IsMoving => m_LastInput.sqrMagnitude > 0.01f;
-
+        
         public void EnableLockCursor(bool enable) => LockCursor = enable;
 
         /// Report the available input axes to the input axis controller.
@@ -76,11 +49,33 @@ namespace Unity.Cinemachine.Samples
             axes.Add(new () { DrivenAxis = () => ref Jump, Name = "Jump" });
             axes.Add(new () { DrivenAxis = () => ref Sprint, Name = "Sprint" });
         }
+    }
 
-        void Start()
-        {
-            TryGetComponent(out m_Controller);
-        }
+    public class SimplePlayerController : SimplePlayerControllerBase
+    {
+        public float Damping = 0.5f;
+        public bool Strafe = false;
+
+        public enum ForwardModes { Camera, Player, World };
+        public ForwardModes InputForward = ForwardModes.Camera;
+        public enum UpModes { Player, World };
+        public UpModes UpMode = UpModes.World;
+
+        [Tooltip("Override the main camera. Useful for split screen games.")]
+        public Camera CameraOverride;
+        
+        Vector3 m_CurrentVelocityXZ;
+        Vector3 m_LastInput;
+        float m_CurrentVelocityY;
+        bool m_IsSprinting;
+        bool m_IsJumping;
+        CharacterController m_Controller; // optional
+
+        public bool IsSprinting => m_IsSprinting;
+        public bool IsJumping => m_IsJumping;
+        public bool IsMoving => m_LastInput.sqrMagnitude > 0.01f;
+
+        void Start() => TryGetComponent(out m_Controller);
 
         private void OnEnable()
         {
@@ -137,7 +132,13 @@ namespace Unity.Cinemachine.Samples
                 transform.rotation = Quaternion.Slerp(qA, qB, Damper.Damp(1, damping, Time.deltaTime));
             }
 
-            PostUpdate?.Invoke();
+            if (PostUpdate != null)
+            {
+                // Get local-space velocity
+                var vel = Quaternion.Inverse(transform.rotation) * m_CurrentVelocityXZ;
+                vel.y = m_CurrentVelocityY;
+                PostUpdate(vel, m_IsSprinting ? JumpSpeed / SprintJumpSpeed : 1);
+            }
         }
 
         Vector3 UpDirection => UpMode == UpModes.World ? Vector3.up : transform.up;
@@ -148,7 +149,7 @@ namespace Unity.Cinemachine.Samples
             var up = UpDirection;
             var fwd = InputForward switch
             {
-                ForwardModes.Camera => Camera.main.transform.forward,
+                ForwardModes.Camera => CameraOverride == null? Camera.main.transform.forward : CameraOverride.transform.forward,
                 ForwardModes.Player => transform.forward,
                 _ => Vector3.forward,
             };
@@ -164,7 +165,7 @@ namespace Unity.Cinemachine.Samples
 
         bool ProcessJump()
         {
-            const float kGravity = -9.8f;
+            const float kGravity = -10;
             bool justLanded = false;
             m_CurrentVelocityY += kGravity * Time.deltaTime;
             if (!m_IsJumping && Jump.Value > 0.01f)
@@ -227,9 +228,9 @@ namespace Unity.Cinemachine.Samples
 
         float GetDistanceFromGround(Vector3 pos, Vector3 up, float max)
         {
-            float hExtraHeight = 2;
-            if (Physics.Raycast(pos + up * hExtraHeight, -up, out var hit, max + hExtraHeight, LayerMask.GetMask("Default")))
-                return hit.distance - hExtraHeight; 
+            float kExtraHeight = 2;
+            if (Physics.Raycast(pos + up * kExtraHeight, -up, out var hit, max + kExtraHeight, LayerMask.GetMask("Default")))
+                return hit.distance - kExtraHeight; 
             return max + 1;
         }
     }