+++ com.unity.cinemachine/CHANGELOG.md
@@ -15,10 +15,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - InputAxisController has the option to suppress input while the attached camera is blending.
 - Added CinemachineCameraEvents and CinemachineBrainEvents behaviours for event processing.
 - Added BlendFinished and CameraDeactivated events.
+- Add split screen sample for the input system.
 - Samples UI works with both built-in Input System.
 - Timeline: Added Track Priority field in CinemachineTrack to control track precedence when tracks are contained in nested timelines.
+- New 2D platformer sample showing custom camera manager.
 
 ### Changed
+- Minimum Unity version is now 2022.2.15f1.
 - All namespaces changed from "Cinemachine" to "Unity.Cinemachine".
 - "Cinemachine.Utility" namespace folded into to "Unity.Cinemachine".
 - CinemachineBlendListCamera has been renamed to CinemachineSequencerCamera.